ec2-monitoring-agent/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.31.0
boto3==1.34.100
botocore>=1.34.100
python-dotenv==1.0.1
APScheduler==3.10.4
